Stories/ecer 1186

+++ src/ECER.Clients.RegistryPortal/ecer.clients.registryportal.client/src/components/FileUploader.vue
@@ -21,6 +21,11 @@
           @delete-file="removeFile"
         ></UploadFileItem>
       </v-list>
+      <v-input
+        :model-value="userFiles"
+        :hide-details="'auto'"
+        :rules="[!fileErrors, !filesInProgress, !fileTooLarge, !allFilesTooLarge, !duplicateFileName, ...rules]"
+      />
     </v-col>
   </v-row>
 </template>
@@ -28,7 +33,7 @@
 <script lang="ts">
 import type { AxiosProgressEvent } from "axios";
 import { v4 as uuidv4 } from "uuid";
-import { defineComponent } from "vue";
+import { defineComponent, type PropType } from "vue";
 
 import { deleteFile, uploadFile } from "@/api/file";
 import Alert from "@/components/Alert.vue";
@@ -40,12 +45,30 @@ interface FileUploaderData {
   selectedFiles: FileItem[];
   showErrorBanner: boolean;
   errorBannerMessage: string;
+  firstLoad: boolean;
 }
 
 export default defineComponent({
   name: "FileUploader",
   components: { UploadFileItem, Alert },
-  emits: ["update:files"],
+  props: {
+    userFiles: {
+      type: Array as PropType<FileItem[]>,
+      required: false,
+      default: () => [],
+    },
+    deleteFileFromTempWhenRemoved: {
+      type: Boolean,
+      required: false,
+      default: true,
+    },
+    rules: {
+      type: Array as PropType<any[]>,
+      required: false,
+      default: () => [],
+    },
+  },
+  emits: ["update:files", "delete:file"],
   async setup() {
     const alertStore = useAlertStore();
     const maxNumberOfFiles = 5;
@@ -64,15 +87,52 @@ export default defineComponent({
       selectedFiles: [],
       showErrorBanner: false,
       errorBannerMessage: "",
+      firstLoad: true,
     };
+  },
+  computed: {
+    fileErrors() {
+      return this.selectedFiles.some((file) => file.fileErrors.length !== 0);
+    },
+    filesInProgress() {
+      return this.selectedFiles.some((file) => file.progress < 101);
+    },
+    fileTooLarge() {
+      return this.selectedFiles.some((file) => file.file.size > this.maxFileSizeInBytes);
+    },
+    allFilesTooLarge() {
+      const totalSize = this.selectedFiles?.reduce((acc, sf) => acc + sf.file.size, 0);
+      if (totalSize > this.maxFileSizeInBytes) {
+        return true;
+      }
+      return totalSize > this.maxFileSizeInBytes;
+    },
+    duplicateFileName() {
+      const fileNameMap = new Set();
+      for (const file of this.selectedFiles) {
+        if (fileNameMap.has(file.file.name)) {
+          return true; //there is a duplicate stop checking and return true
+        }
+        fileNameMap.add(file.file.name);
+      }
+
+      return false;
+    },
   },
   watch: {
     selectedFiles: {
       handler() {
-        this.updateEmit(); // Emit updates whenever selectedFiles changes
+        if (!this.firstLoad) {
+          this.updateEmit(); // Emit updates whenever selectedFiles changes
+        } else {
+          this.firstLoad = false;
+        }
       },
       deep: true,
     },
+  },
+  mounted() {
+    this.selectedFiles = [...this.userFiles];
   },
   methods: {
     handleFileUpload(event: Event) {
@@ -93,7 +153,15 @@ export default defineComponent({
             fileErrors.push("A file with the same name already exists.");
           }
           const fileId = uuidv4(); // Generate a unique file ID using uuid
-          const selectedFile = { file, progress: 0, fileId, fileErrors };
+          const selectedFile: FileItem = {
+            file,
+            fileSize: file.size,
+            fileName: file.name,
+            progress: 0,
+            fileId,
+            fileErrors,
+            storageFolder: "temporary",
+          };
           this.selectedFiles.push(selectedFile);
           if (this.selectedFiles.length > 1) {
             const totalSize = this.selectedFiles.reduce((acc, sf) => acc + sf.file.size, 0);
@@ -151,9 +219,10 @@ export default defineComponent({
     async removeFile(selectedFile: FileItem) {
       try {
         this.selectedFiles = this.selectedFiles.filter((f: FileItem) => f.fileId !== selectedFile.fileId);
-        if (!(selectedFile.fileErrors.length > 0)) {
+        if (!(selectedFile.fileErrors.length > 0) && selectedFile.storageFolder === "temporary" && this.deleteFileFromTempWhenRemoved) {
           await deleteFile(selectedFile.fileId);
         }
+        this.$emit("delete:file", selectedFile);
         this.showErrorBanner = false;
       } catch (error) {
         this.alertStore.setFailureAlert("An error occurred during file deletion.");
